Change the complexity of a fit_estimate_partition

Change the complexity level of the partition and re-estimate cell statistics. If you have a minimal estimated_partition then you need to pass in the other params.

change_complexity(
  fit,
  y,
  X,
  d = NULL,
  partition_i,
  index_tr = fit$index_tr,
  split_seq = fit$split_seq,
  est_plan = fit$est_plan
)

Arguments

fit
estimated_partition

y
Nx1 matrix of outcome (label/target) data. With multiple core estimates see Details below.

X
NxK matrix of features (covariates). With multiple core estimates see Details below.

d
(Optional) NxP matrix (with colnames) of treatment data. If all equally important they should be normalized to have the same variance. With multiple core estimates see Details below.

partition_i
partition_i - 1 is the last include in split_seq included in new partition

index_tr
Split between train and estimate samples (default is to get from fit)

split_seq
sequential list of splits (default is to get from fit)

est_plan
EstimatorPlan.

Value

updated estimated_partition

Details

Note: doesn't update the importance weights

Multiple estimates

With multiple core estimates (M) there are 3 options (the first two have the same sample across treatment effects).

DS.MULTI_SAMPLE: Multiple pairs of (Y_m,W_m). y,X,d are then lists of length M. Each element then has the typical size The N_m may differ across m. The number of columns of X will be the same across m.

DS.MULTI_D: Multiple treatments and a single outcome. d is then a NxM matrix.

DS.MULTI_Y: A single treatment and multiple outcomes. y is then a NXM matrix.